requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-cors-headers==2.5.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.8
oauthlib==3.0.0
Pillow==5.4.1
pkg-resources==0.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
urllib3==1.24.1

